+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ## Unreleased
 
+## 1.5.0
+
 - Add remove button to link tooltip
 - Fix tabbing through the toolbar in Firefox - [PR](https://github.com/alphagov/govspeak-visual-editor/pull/123)
 - Add skipped heading level linting - [PR #125](https://github.com/alphagov/govspeak-visual-editor/pull/125)